Add files via upload
This is the simplecpp package that I've used for writing some of my programs in C++, as was demanded by the course. However, this package is only truly needed for the programs where the turtle is used. Otherwise there are plenty of "pure" C++ files too which run as it is.